services/web/requirements.txt
psycopg2-binary==2.8.6
gunicorn==20.0.4
SQLAlchemy~=1.3.22
APScheduler~=3.7.0
eventlet~=0.30.0
qrcode~=6.1
selenium~=3.141.0
xxhash~=2.0.0
PyJWT~=2.0.1
alembic~=1.5.2
Werkzeug==1.0.1
requests==2.25.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Flask-Limiter
Flask-Admin
Flask-AdminLTE3
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.5